Floatify, an attempt to fix interpolation

+++ include/GLWidget.hpp
@@ -25,8 +25,8 @@ public slots:
     void sigShowRepeated(bool show);
 
     void sigFindInRadius();
-    void sigSetRadius(double r);
-    void sigSetH(double h);
+    void sigSetRadius(float r);
+    void sigSetH(float h);
     void sigSetK(int k);
     void sigSetN(int n);
     void sigSetM(int m);
