+++ include/Grid3D.hpp
@@ -25,12 +25,12 @@ public:
 
     virtual void approximateWLS(VertexList& resultList);
     virtual void draw(bool useAlpha);
-	virtual double getInterpolatedFunctionValue(float x,float y,float z,int valueType);
-	virtual double getImplicitFunctionValueWorldCoordinates(float x,float y,float z,int valueType);
+	virtual float getInterpolatedFunctionValue(float x,float y,float z,int valueType);
+	virtual float getImplicitFunctionValueWorldCoordinates(float x,float y,float z,int valueType);
 
 	int getIndex(int idX,int idY,int idZ);
 	VertexPtr getVertex(int idX,int idY,int idZ);
-	double getVertexValue(int idX,int idY,int idZ,int valueType);
+	float getVertexValue(int idX,int idY,int idZ,int valueType);
     NormalPtr interpolateNormal(VertexList vrtxList) const;
 	size_t getDimZ() {return m_dimZ;};
